% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Beaf32_Peaks.gnr.R
\docType{data}
\name{Beaf32_Peaks.gnr}
\alias{Beaf32_Peaks.gnr}
\title{D.melanogaster Beaf-32 ChIP-seq.}
\format{
An object of class GRanges.
}
\usage{
data(Beaf32_Peaks.gnr)
}
\description{
Drosophila Melanogaster Beaf32 peaks on 2L and 2R chromosomes.
}
\examples{
data(Beaf32_Peaks.gnr)
Beaf32_Peaks.gnr

}
\keyword{datasets}
