% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxBlur.R
\name{BoxBlur}
\alias{BoxBlur}
\title{Blur a matrix.}
\usage{
BoxBlur(mtx, boxKernel = NULL, kernSize = NULL, stdev = 1)
}
\arguments{
\item{mtx}{<matrix>: Numerical matrix.}

\item{boxKernel}{<numeric>: The numerical vector for kernel. If NULL apply a GaussBox (see 'GaussBox' function) (Default NULL)}

\item{kernSize}{<numeric>: If boxKernel is NULL, size of kernel for 'GaussBox' function. (Default NULL)}

\item{stdev}{<numeric>: If boxKernel is NULL, standard deviation parameter for 'GaussBox' function. (Default NULL)}
}
\value{
Blurred matrix.
}
\description{
Blur a matrix with a one dimensional kernel.
}
\details{
BoxBlur
}
\examples{
set.seed(981643)
mtx <- rnorm(10000, 50, 10)**3 |> matrix(100, 100)
heatmap(mtx, Rowv = NA, Colv = NA)
heatmap(BoxBlur(mtx), Rowv = NA, Colv = NA)

}
\keyword{internal}
