% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutHiC.R
\name{CutHiC}
\alias{CutHiC}
\title{Cut HiC map in chunks.}
\usage{
CutHiC(megaHic, verbose = FALSE)
}
\arguments{
\item{megaHic}{\if{html}{\out{<contactMatrix>}}: The HiC megamap.}

\item{verbose}{\if{html}{\out{<logical>}}: If TRUE show the progression in console. (Default FALSE)}
}
\value{
A matrices list.
}
\description{
Cut a mega contactMatrix (join of multiple chromosomic maps) inq a list of contactMatrix.
}
\details{
CutHiC
}
\examples{
data(HiC_Ctrl.cmx_lst)
Mega_Ctrl.cmx <- JoinHiC(HiC_Ctrl.cmx_lst)
CutHiC(Mega_Ctrl.cmx)

}
