% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gauss.R
\name{Gauss}
\alias{Gauss}
\title{Gaussian formula.}
\usage{
Gauss(x = NULL, y = NULL, stdev = 1, mu = 0)
}
\arguments{
\item{x}{\if{html}{\out{<numeric>}}: x value.}

\item{y}{\if{html}{\out{<numeric>}}: y value for 2 dimensional gaussian.}

\item{stdev}{\if{html}{\out{<numeric>}}: Standard deviation parameter of the gaussian. (Default 1)}

\item{mu}{\if{html}{\out{<numeric>}}: Mean deviation parameter of the gaussian. (Default 0)}
}
\value{
Result of Gaussian formula
}
\description{
Gaussian formula in 1 or 2 dimension.
}
\details{
Gauss
}
\examples{
Gauss(x = 1)
Gauss(x = 1, y = 2)

}
\keyword{internal}
