% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussBox.R
\name{GaussBox}
\alias{GaussBox}
\title{One dimension Gaussian kernel.}
\usage{
GaussBox(stdev = 1, kernSize = NULL, kernScale = "1")
}
\arguments{
\item{stdev}{<numeric>: Standard deviation parameter of the gaussian. (Default 1)}

\item{kernSize}{<numeric>: Kernel size. If NULL size is 1+4*stdev (Default NULL)}

\item{kernScale}{<character>: Scaling kind of box. If "1" sum of kernel equal 1. If "int" Minimal value of kernel is 1 and all entry are integer. If "none", kernel is not scale. (Default "1")}
}
\value{
numerical vector.
}
\description{
One dimension Gaussian kernel.
}
\details{
GaussBox
}
\examples{
GaussBox(stdev = 5, kernScale = "none")
GaussBox(kernScale = "1")
GaussBox(kernScale = "int")

}
\keyword{internal}
