% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GenomicSystem}
\alias{GenomicSystem}
\title{Base pairs convertion.}
\usage{
GenomicSystem(x, digits = 3)
}
\arguments{
\item{x}{\if{html}{\out{<character or numeric>}}: The number
to convert or string to convert.}

\item{digits}{\if{html}{\out{<integer>}}: The number of significant
digits to be used. See \code{\link[=signif]{signif()}} for more informations. (Default 3)}
}
\value{
The converted number or string.
}
\description{
Convert numbers of base into string with order of magnitude
(Kbp, Mbp, Gbp) and vice versa.
}
\details{
GenomicSystem
}
\examples{
GenomicSystem(1540, 3)
GenomicSystem(1540, 2)
GenomicSystem("1Mbp")
GenomicSystem("1Kbp")
GenomicSystem("1k")

}
