% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{GetFileName}
\alias{GetFileName}
\title{Get file name.}
\usage{
GetFileName(file = NULL, extension = FALSE)
}
\arguments{
\item{file}{\if{html}{\out{<character>}}: The path to the file.}

\item{extension}{\if{html}{\out{<logical>}}: Whether the file extension should be returned
with the file name. (Default FALSE)}
}
\value{
A character string.
}
\description{
Function as \code{basename()} with the option to not return the file
extension.
}
\details{
GetFileName
}
\examples{
filePath.pth ="my/path/to/my/file.txt"
GetFileName(file=filePath.pth, extension=FALSE)
GetFileName(file=filePath.pth, extension=TRUE)
}
\keyword{internal}
