% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInfos.R
\name{getInfos}
\alias{getInfos}
\title{get some basic information about your hic file}
\usage{
getInfos(file = NULL, printInfos = TRUE, returnInfos = FALSE)
}
\arguments{
\item{file}{\if{html}{\out{<character>}} path to your file.}

\item{printInfos}{\if{html}{\out{<logical>}} print info on console?
(Default: TRUE)}

\item{returnInfos}{\if{html}{\out{<logical>}} return info? (Default: FALSE)}
}
\value{
list of characters if \code{returnInfos} is TRUE.
}
\description{
getInfos
}
\examples{
h5_path <- system.file("extdata",
    "Control_HIC_10k_2L.h5",
    package = "HicAggR", mustWork = TRUE
)
getInfos(h5_path)

}
