% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HiC_Ctrl.cmx_lst.R
\docType{data}
\name{HiC_Ctrl.cmx_lst}
\alias{HiC_Ctrl.cmx_lst}
\title{In situ Hi-C control.}
\format{
A a list of ContactMatrix objects. Each element correspond to the interaction matrix of two chromosomes.
}
\usage{
data(HiC_Ctrl.cmx_lst)
}
\description{
In situ Hi-C on non-heat treated S2 cells (Drosophila Melanogaster) with MboI on chromosome 2R and 2L download from [4DN](https://data.4dnucleome.org/experiment-set-replicates/4DNESFOADERB/) portal (Ray J, Munn PR, et al., 2019). This data is the result of the HicAggR::ImportHiC function.
}
\examples{
data(HiC_Ctrl.cmx_lst)
HiC_Ctrl.cmx_lst

}
\keyword{datasets}
