% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Hsl2Hex}
\alias{Hsl2Hex}
\title{Convert HSL to Hex.}
\usage{
Hsl2Hex(hslColor = NULL, alpha = FALSE)
}
\arguments{
\item{hslColor}{\if{html}{\out{<charcater>}}: A vector of the color's HSL code.}

\item{alpha}{\if{html}{\out{<logical>}}: Whether the alpha layer should be returned.
(Default FALSE)}
}
\value{
A character of the color's hexadecimal code.
}
\description{
Convert a color in HSL (Hue,Saturation,Light) to
hexadecimal format.
}
\details{
Hsl2Hex
}
\examples{
Hsl2Hex(c(43.8, 0.873, 0.492, 0.498), alpha = TRUE)

}
\keyword{internal}
