% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Hsl2Rgb}
\alias{Hsl2Rgb}
\title{Convert HSL to RGB.}
\usage{
Hsl2Rgb(hslColor = NULL, alpha = FALSE)
}
\arguments{
\item{hslColor}{\if{html}{\out{<charcater>}}: A vector of the color's HSL code.}

\item{alpha}{\if{html}{\out{<logical>}}: Whether the alpha layer should be returned.
(Default FALSE)}
}
\value{
An integer vector of the color's RGB code.
}
\description{
Convert a color in HSl (Hue,Saturation,Light) format to RGB.
}
\details{
Hsl2Rgb
}
\examples{
Hsl2Rgb(c(43.8, 0.873, 0.492, 0.498), alpha = TRUE)

}
\keyword{internal}
