% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hue.R
\name{Hue}
\alias{Hue}
\title{Hue palette.}
\usage{
Hue(
  paletteLength = 9,
  rotation = NULL,
  hueRange = c(0, 360),
  saturation = 0.65,
  lightness = 0.65,
  alphaValue = 1,
  alpha = FALSE
)
}
\arguments{
\item{paletteLength}{<numeric>: Color number.}

\item{rotation}{<numeric>: If positive, rotates clockwise in the color space, reversing if the number is negative. If is NULL compute rotation according to hueRange parameter. (Default NULL)}

\item{hueRange}{<numeric>: Degree range in color space between 0 and 360. (Default c(0,360))}

\item{saturation}{<numeric>: Saturation value between 0 and 1. (Default 0.65)}

\item{lightness}{<numeric>: Lightness value between 0 and 1. (Default 0.65)}

\item{alphaValue}{<numeric>: Opacity value between 0 and 1. (Default 1)}

\item{alpha}{<logical>: Whether the alpha layer should be returned. (Default FALSE)}
}
\value{
A vector of color.
}
\description{
Create an Hue palette.
}
\details{
Hue
}
\examples{
Hue(paletteLength = 9)

}
