% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportLoops.R
\name{ImportLoops}
\alias{ImportLoops}
\title{ImportLoops: import called loops in .bedpe format to use in HicAggR}
\usage{
ImportLoops(
  file_bedpe = NULL,
  genomicConstraint = NULL,
  discard_trans = FALSE,
  chromSizes = NULL,
  binSize = NULL,
  minDist = NULL,
  maxDist = NULL,
  verbose = FALSE,
  cores = 1
)
}
\arguments{
\item{file_bedpe}{bedpe file path (Default NULL)}

\item{genomicConstraint}{\if{html}{\out{<GRanges>}}: GRanges object of
constraint regions. If NULL chromosomes in chromSizes are used as
constraints (Default NULL).}

\item{discard_trans}{\if{html}{\out{<logical>}}: If TRUE discard loops where
anchor and bait are in different genomic constraint elements,
either different TADs or chromosomes, if \code{genomicConstraint = NULL}.
(Default FALSE)}

\item{chromSizes}{<data.frame>: A data.frame containing
chromosomes names and lengths in base pairs (see example). (Default NULL)}

\item{binSize}{\if{html}{\out{<integer>}}: Bin size in bp - corresponds to
matrix resolution.}

\item{minDist}{\if{html}{\out{<numeric>}}: Minimal distance between anchors
and baits. (Default NULL)}

\item{maxDist}{\if{html}{\out{<numeric>}}: Maximal distance between anchors
and baits. (Default NULL)}

\item{verbose}{\if{html}{\out{<logical>}}: Show the progression in console?
(Default FALSE)}

\item{cores}{\if{html}{\out{<integer>}}: Number of cores to use. (Default 1)}
}
\value{
A GInteractions object.
}
\description{
Imports bedpe file formats as GInteractions object usable
to perform submatrix extraction with \code{ExtractSubmatrix()}
}
\examples{

bedpe_path <- system.file("extdata",
    "postprocessed_pixels_5000.bedpe",
    package = "HicAggR", mustWork = TRUE
)
loops <- ImportLoops(
    file_bedpe = bedpe_path,
    chromSizes = data.frame(c("1","2"),
        c(249250621,243199373)),
    binSize = 5000, minDist = 105000
)

}
