% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsHsl.R
\name{IsHsl}
\alias{IsHsl}
\title{Check HSL color format.}
\usage{
IsHsl(color.col = NULL)
}
\arguments{
\item{color.col}{<character or numeric>: A color.}
}
\value{
A logical.
}
\description{
Check if a color is in HSL color format.
}
\details{
IsHsl
}
\examples{
IsHsl("red")
IsHsl("#FFFFFF")
IsHsl(c(125, 125, 125))
IsHsl(c(43.8, 0.873, 0.492))

}
\keyword{internal}
