% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoinHiC.R
\name{JoinHiC}
\alias{JoinHiC}
\title{Merge HiC chunk.}
\usage{
JoinHiC(hicLst)
}
\arguments{
\item{hicLst}{<List\link[InteractionSet:ContactMatrix-class]{ContactMatrix}>:
The HiC maps list.}
}
\value{
A ContactMatrix.
}
\description{
Create mega contactMatrix from a list of contactMatrix.
}
\details{
JoinHiC
}
\examples{
data(HiC_Ctrl.cmx_lst)
Mega_Ctrl.cmx <- JoinHiC(HiC_Ctrl.cmx_lst)

}
