% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeParallelParam.R
\name{MakeParallelParam}
\alias{MakeParallelParam}
\title{Configure parallel parameters.}
\usage{
MakeParallelParam(cores.num = 1, verbose.bln = FALSE)
}
\arguments{
\item{cores.num}{<numerical> : An integer to specify the number of cores. (Default 1)}

\item{verbose.bln}{<logical>: A logical value. If TRUE show the progression in console. (Default TRUE)}
}
\value{
Parrallel parameter according number of cores and OS to use with BiocParallel package.
}
\description{
Create BiocParallel parameter according to OS.
}
\details{
MakeParallelParam
}
\examples{
multicoreParam <- MakeParallelParam(2)
BiocParallel::bplapply(BPPARAM = multicoreParam, seq_len(3), sqrt)

}
\keyword{internal}
