% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeltSpm.R
\name{MeltSpm}
\alias{MeltSpm}
\title{Coerce matrix in tibble.}
\usage{
MeltSpm(spMtx = NULL)
}
\arguments{
\item{spMtx}{<dgCMatrix or dgCMatrix coercible>: A matrix.}
}
\value{
A tibble.
}
\description{
Coerce a sparse matrix M in tibble where columns: i is row index, j is column index and x the value M[i,j].
}
\details{
MeltSpm
}
\examples{
i <- c(1, 1, 2, 2, 3, 3, 4, 4, 4, 4)
j <- c(1, 4, 2, 5, 1, 4, 2, 3, 4, 5)
x <- c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)
spMtx <- Matrix::sparseMatrix(i = i, j = j, x = x, dims = c(5, 5))
spMtx
meltedMat.tbl <- MeltSpm(spMtx)
meltedMat.tbl[order(meltedMat.tbl$i), ]

}
\keyword{internal}
