% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotIn.R
\name{NotIn}
\alias{NotIn}
\title{Exclusion binary operator.}
\usage{
NotIn(lhs, rhs)
}
\arguments{
\item{lhs}{<vector or NULL>: Values to be compared against rhs}

\item{rhs}{<vector or NULL>: Values to be compared against lhs}
}
\value{
A boolean.
}
\description{
Binary operator, inverse to \%in\%.
}
\details{
NotIn
}
\examples{
"A" |> NotIn(c("A", "B", "C"))
"A" |> NotIn(c("B", "C", "D"))
NotIn("A", c("A", "B", "C"))
NotIn("A", c("B", "C", "D"))

}
\keyword{internal}
