% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{PadMtx}
\alias{PadMtx}
\title{Add a value around a matrix.}
\usage{
PadMtx(
  mtx = NULL,
  padSize = 1,
  val = 0,
  side = c("top", "bot", "right", "left")
)
}
\arguments{
\item{mtx}{\if{html}{\out{<matrix>}}: Numerical matrix.}

\item{padSize}{\if{html}{\out{<numeric>}}: Number of columns or rows to add. (Default 1)}

\item{val}{\if{html}{\out{<numeric>}}: Value to add. If Null create mirror of choosen sides.
(Default 0)}

\item{side}{\if{html}{\out{<character>}}: Side to pad, must be one or some of 'top','bot',
'right' or 'left'. (Default c('top','bot','right','left') )}
}
\value{
A matrix.
}
\description{
Add a value around a matrix.
}
\details{
PadMtx
}
\examples{
mtx <- matrix(seq_len(25), 5, 5)
PadMtx(
    mtx = mtx,
    padSize = 1,
    val = 0,
    side = c("top", "bot", "right", "left")
)
PadMtx(
    mtx = mtx,
    padSize = 1,
    val = NULL,
    side = c("top", "bot", "right", "left")
)
PadMtx(
    mtx = mtx,
    padSize = 1,
    val = 0,
    side = c("right", "left")
)
PadMtx(
    mtx = mtx,
    padSize = 1,
    val = 0,
    side = c("top")
)

}
\keyword{internal}
