% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotAPA.r
\name{PlotAPA}
\alias{PlotAPA}
\title{Draw aggregation plot.}
\usage{
PlotAPA(
  aggregatedMtx = NULL,
  trim = 0,
  colMin = NULL,
  colMid = NULL,
  colMax = NULL,
  colMinCond = NULL,
  colMaxCond = NULL
)
}
\arguments{
\item{aggregatedMtx}{\if{html}{\out{<matrix>}}: The aggregated matrix.}

\item{trim}{\if{html}{\out{<numeric>}}: A number between 0 and 100 thaht give the percentage of triming in matrices.}

\item{colMin}{\if{html}{\out{<matrix>}}: The minimal value in color scale. If Null automaticaly find.}

\item{colMid}{\if{html}{\out{<matrix>}}: The middle value in color scale. If Null automaticaly find.}

\item{colMax}{\if{html}{\out{<matrix>}}: The mximal value in color scale. If Null automaticaly find.}

\item{colMinCond}{\if{html}{\out{<matrix>}}: Avalaible for plotting differantial aggregation. The minimal value in color scale in the classsical aggregation plot. If Null automaticaly find.}

\item{colMaxCond}{\if{html}{\out{<matrix>}}: Avalaible for plotting differantial aggregation. The maxiaml value in color scale in the classsical aggregation plot. If Null automaticaly find.}
}
\value{
None
}
\description{
Draw aggregation plot from aggregation matrices.
}
\details{
PlotAPA
}
\examples{
library(GenomicED)
data(aggreg.mtx)


PlotAPA(
    aggregatedMtx                  = aggreg.mtx,
    trim             = 20,
    colMin          = -2,
    colMid               = 0,
    colMax          = 2,
    colMinCond = 0,
    colMaxCond = 2
)


}
