% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plus.R
\name{Plus}
\alias{Plus}
\title{Sum by removing NA.}
\usage{
Plus(x)
}
\arguments{
\item{x}{<numerical>: A numerical vector}
}
\value{
The sum of numbers.
}
\description{
Perform sum by removing NA. If all values are NA return NA instead 0.
}
\details{
Plus
}
\examples{
Plus(c(1, 2, 3))
sum(c(1, 2, 3), na.rm = TRUE)
sum(c(1, 2, 3))
Plus(c(1, 2, NA))
sum(c(1, 2, NA), na.rm = TRUE)
sum(c(1, 2, NA))
Plus(c(NA, NA, NA))
sum(c(NA, NA, NA), na.rm = TRUE)
sum(c(NA, NA, NA))

}
\keyword{internal}
