% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ReduceRun}
\alias{ReduceRun}
\title{Apply a function over two RLE}
\usage{
ReduceRun(firstRle, secondRle, reduceMethod = "paste", ...)
}
\arguments{
\item{firstRle}{\if{html}{\out{<rle or Rle>}}: First rle.}

\item{secondRle}{\if{html}{\out{<rle or Rle>}}>: Second rle.}

\item{reduceMethod}{\if{html}{\out{<character>}}: Name of a function to apply
e.g paste, sum, mean.}

\item{...}{<...>: Other parameter for the reduce function.}
}
\value{
Reduced Rle
}
\description{
Apply a function on the values over two RLE and return one RLE.
}
\details{
ReduceRun
}
\keyword{internal}
