% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResizeMatrix.R
\name{ResizeMatrix}
\alias{ResizeMatrix}
\title{Resize a matrix}
\usage{
ResizeMatrix(mtx, newDim = dim(mtx))
}
\arguments{
\item{mtx}{\if{html}{\out{<matrix>}}: A numerical matrix to resize.}

\item{newDim}{\if{html}{\out{<integer>}}: The number of rows and cols in resized matrix.}
}
\value{
Resized matrix.
}
\description{
Resize a numericam matrix in new dimension.
}
\details{
ResizeMatrix
}
\examples{
mtx <- matrix(0, 11, 11)
mtx[which(as.logical(seq_len(11 * 11) \%\% 2))] <- seq_len(ceiling((11 * 11) / 2))
mtx[2, ] <- 100
mtx[, 7] <- 200
mtx
ResizeMatrix(mtx = mtx, newDim = c(7, 7))
ResizeMatrix(mtx = mtx, newDim = c(13, 13))
}
\keyword{internal}
