% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SdThreshold.R
\name{SdThreshold}
\alias{SdThreshold}
\title{Find threshold for outliers based on sd.}
\usage{
SdThreshold(x = NULL, sdThr = 3, tails = "both")
}
\arguments{
\item{x}{\if{html}{\out{<numeric>}}: numeric vector.}

\item{sdThr}{\if{html}{\out{<numeric>}}: number of standard deviation. (Default 3)}

\item{tails}{\if{html}{\out{<character>}}: bounds to return, "lower", "upper" or "both". (Default "both")}
}
\value{
numerical vector of thresholds values for outliers triming
}
\description{
Find threshold for outliers triming based on standard deviation.
}
\details{
SdThreshold
}
\examples{
set.seed(1111)
x <- rnorm(1000)
x <- sort(x)
x
SdThreshold(x, sdThr = 2, tails = "lower")
SdThreshold(x, sdThr = 2, tails = "both")
SdThreshold(x, sdThr = 2, tails = "upper")

}
\keyword{internal}
