% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StrToGRanges.R
\name{StrToGRanges}
\alias{StrToGRanges}
\title{Convert String to GRanges.}
\usage{
StrToGRanges(x.chr_vec)
}
\arguments{
\item{x.chr_vec}{<character>: Strings to convert on GRanges.}
}
\value{
A GRanges object.
}
\description{
Convert ranges describe with string (i.e seqname:start-end:strand) in GRanges object.
}
\details{
StrToGRanges
}
\examples{
StrToGRanges("chr1:1-100:+")
StrToGRanges(c("chr1:1-100:+", "chr2:400-500:-", "chr1:10-50:*"))

}
