% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SwitchMatrix.R
\name{SwitchMatrix}
\alias{SwitchMatrix}
\title{Change values of HiC map.}
\usage{
SwitchMatrix(hic.cmx_lst, matrixKind.chr = c("obs", "norm", "o/e", "exp"))
}
\arguments{
\item{hic.cmx_lst}{<List[contactMatrix]>: The HiC maps list.}

\item{matrixKind.chr}{<character>: The kind of matrix you want.}
}
\value{
A contactMatrix list.
}
\description{
Change values in matrix with observed, balanced, observed/expected or expected values according to what are be done in hic.
}
\details{
SwitchMatrix
}
\examples{
# Data
data(HiC_Ctrl.cmx_lst)

# Preprocess HiC
HiC.cmx_lst <- HiC_Ctrl.cmx_lst |>
    BalanceHiC(
        interaction.type = "cis",
        method.chr = "ICE"
    ) |>
    OverExpectedHiC()

# Switch values in matrix
HiC_Ctrl_Obs.cmx_lst <- SwitchMatrix(HiC.cmx_lst, matrixKind.chr = "obs")
HiC_Ctrl_Norm.cmx_lst <- SwitchMatrix(HiC.cmx_lst, matrixKind.chr = "norm")
HiC_Ctrl_oe.cmx_lst <- SwitchMatrix(HiC.cmx_lst, matrixKind.chr = "o/e")
HiC_Ctrl_exp.cmx_lst <- SwitchMatrix(HiC.cmx_lst, matrixKind.chr = "exp")

}
