% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TADs_Domains.gnr.R
\docType{data}
\name{TADs_Domains.gnr}
\alias{TADs_Domains.gnr}
\title{D.melanogaster TADs.}
\format{
An object of class GRanges.
}
\usage{
data(TADs_Domains.gnr)
}
\description{
Drosophila Melanogaster TADs on chromosome 2R and 2L ([F.Ramirez, 2018](https://doi.org/10.1038/s41467-017-02525-w))
}
\examples{
data(TADs_Domains.gnr)
TADs_Domains.gnr

}
\keyword{datasets}
