% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TransposeList.R
\name{TransposeList}
\alias{TransposeList}
\title{Turns a nested list "inside-out".}
\usage{
TransposeList(var.nlst)
}
\arguments{
\item{var.nlst}{<list[list]>: A nested list to transpose.}
}
\value{
The tranposed nested list.
}
\description{
Turns a nested list "inside-out".
}
\details{
TransposeList
}
\examples{
my_lst <- list(
    first = list("A1", "B1", "C1"),
    second = list("A2", "B2"),
    third = list(NULL, "B3")
)
TransposeList(my_lst)

}
\keyword{internal}
