% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{TrimOutliers}
\alias{TrimOutliers}
\title{Remove outliers.}
\usage{
TrimOutliers(x, thr = SdThreshold(x), clip = FALSE)
}
\arguments{
\item{x}{\if{html}{\out{<numeric>}}: Numeric vector.}

\item{thr}{\if{html}{\out{<numeric>}}: Numeric vector of length 2. first
value is minimal threshold, second value maximal threshold
(Default find threshold based on standarrd deviation.
see \code{SdThreshold} function)}

\item{clip}{\if{html}{\out{<logical>}}: If TRUE the values out of bounds are
replaced with thresholds values. If FALSE the Values out of bound are
replaced with NA (Default FALSE).}
}
\value{
Trimed Numerical vector.
}
\description{
Replace values of a numerical vector that are below a minimal
thresholds and/or above maximal thresholds.
}
\details{
TrimOutliers
}
\keyword{internal}
