% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSS_Peaks.gnr.R
\docType{data}
\name{TSS_Peaks.gnr}
\alias{TSS_Peaks.gnr}
\title{D.melanogaster Transcription starting sites.}
\format{
An object of class GRanges.
}
\usage{
data(TSS_Peaks.gnr)
}
\description{
Data from a CHip Seq experiment
}
\examples{
data(TSS_Peaks.gnr)
TSS_Peaks.gnr

}
\keyword{datasets}
