% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCnorm.R
\name{VCnorm}
\alias{VCnorm}
\title{Compute Vanilla Count Correction.}
\usage{
VCnorm(hic.cmx = NULL, qtlTh.num = 0.15, sqrt.bln = TRUE)
}
\arguments{
\item{hic.cmx}{<contactMatrix>: The HiC maps chunk to normalize.}

\item{qtlTh.num}{<numerical>: The threshold quantile below which the bins will be ignored. (Default 0.15)}

\item{sqrt.bln}{<logical>: Whether the square root must be apply. (Default TRUE)}
}
\value{
A matrices list.
}
\description{
Compute Vanilla Count or Vanilla Count square root correction normalization on hic maps.
}
\details{
VCnorm
}
\examples{
# Data
data(HiC_Ctrl.cmx_lst)

HiC_Ctrl_VC.cmx <- VCnorm(HiC_Ctrl.cmx_lst[["2L_2L"]])
HiC_Ctrl_VC_SQRT.cmx <- VCnorm(HiC_Ctrl.cmx_lst[["2L_2L"]], sqrt.bln = TRUE)

}
