% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{WrapFunction}
\alias{WrapFunction}
\title{Convert string to function.}
\usage{
WrapFunction(...)
}
\arguments{
\item{...}{\if{html}{\out{<character>}}: A string that could be parse and eval as a function.}
}
\value{
The result of the function or a function.
}
\description{
Wrap a string into a function.
}
\details{
WrapFunction
}
\keyword{internal}
