% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{YlGnBu}
\alias{YlGnBu}
\title{YlGnBu palette.}
\usage{
YlGnBu(
  paletteLength = NULL,
  space = "rgb",
  interpolationMethod = "linear",
  bias = 1
)
}
\arguments{
\item{paletteLength}{\if{html}{\out{<numeric>}}: Color number.}

\item{space}{\if{html}{\out{<numeric>}}: A character string; interpolation
in RGB or CIE Lab color spaces. See ?grDevices::colorRamp for more details.
(Default "rgb")}

\item{interpolationMethod}{\if{html}{\out{<numeric>}}: Use spline or linear
interpolation. See ?grDevices::colorRamp for more details.
(Default "linear")}

\item{bias}{\if{html}{\out{<numeric>}}: A positive number. Higher values give
more widely spaced colors at the high end. See \code{?grDevices::colorRamp}
for more details.
(Default 1)}
}
\value{
A vector of color.
}
\description{
Create a YlGnBu palette.
}
\details{
YlGnBu
}
\examples{
YlGnBu(9)

}
