% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddAttr.R
\name{AddAttr}
\alias{AddAttr}
\title{Add list as attributes.}
\usage{
AddAttr(var.any = NULL, attribute.lst = NULL, overwrite.bln = FALSE)
}
\arguments{
\item{var.any}{<any>: An object to which attributes are to be added.}

\item{attribute.lst}{<list>: A named list of new attributes.}

\item{overwrite.bln}{<logical>: Whether an overwrite is required on attributes with the same name.(Default FALSE)}
}
\value{
The object with new attributes.
}
\description{
Add list as attributes to any object with or without overwrite.
}
\details{
AddAttr
}
\examples{
x <- seq_len(10)
x <- AddAttr(x, list(dim = c(2, 5)))
x
x <- AddAttr(x, list(dim = c(5, 2)))
x
x <- AddAttr(x, list(dim = c(5, 2)), overwrite.bln = TRUE)
x

}
\keyword{internal}
