% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BalanceHiC.R
\name{BalanceHiC}
\alias{BalanceHiC}
\title{Compute HiC matrix-balancing.}
\usage{
BalanceHiC(
  hic.cmx_lst,
  method.chr = "ICE",
  interaction.type = NULL,
  maxIter.num = 50,
  qtlTh.num = 0.15,
  cores.num = 1,
  verbose.bln = FALSE
)
}
\arguments{
\item{hic.cmx_lst}{<List[contactMatrix]>: The HiC maps list.}

\item{method.chr}{<character> : The kind of normalization method. One of "ICE", "VC" or "VC_SQRT" (Default "ICE")}

\item{interaction.type}{<character> : "cis", "trans", c("cis", "trans"), "all". If NULL normalization is apply on cis contactMatrix then trans contactMatrix (equivalent to c("cis", "trans")). If is "all", normalization is apply on all contactMatrix at once. (Default NULL)}

\item{maxIter.num}{<numerical>: The maximum iteration number.}

\item{qtlTh.num}{<numerical>: The threshold quantile below which the bins will be ignored. (Default 0.15)}

\item{cores.num}{<numerical> : Number of cores to be used. (Default 1)}

\item{verbose.bln}{<logical>: If TRUE show the progression in console. (Default FALSE)}
}
\value{
A matrices list.
}
\description{
Apply a matrix-balancing normalization method to a list of contacts matrix.
}
\details{
BalanceHiC
}
\examples{
data(HiC_Ctrl.cmx_lst)

HiC_Ctrl_ICE.cmx_lst <- BalanceHiC(HiC_Ctrl.cmx_lst,
    interaction.type = "cis",
    method.chr = "ICE"
)

HiC_Ctrl_VC.cmx_lst <- BalanceHiC(HiC_Ctrl.cmx_lst,
    interaction.type = c("cis", "trans"),
    method.chr = "VC"
)

HiC_Ctrl_VC_SQRT.cmx_lst <- BalanceHiC(HiC_Ctrl.cmx_lst,
    interaction.type = "all",
    method.chr = "VC_SQRT"
)

}
