% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BindFillRows.R
\name{BindFillRows}
\alias{BindFillRows}
\title{Data frames row binding.}
\usage{
BindFillRows(data.lst_df, ...)
}
\arguments{
\item{data.lst_df}{<data.frames or list[data.frame]>: Data frames to bind or list of data.frames. If is a data.frame create a list with arguments `data.lst_df` and `...`, else `...` are ignored.}

\item{...}{<data.frames or list[data.frame]>: Data frames to bind or list of data.frames.}
}
\value{
The binded data frame
}
\description{
Bind data frames by rows after filling missing columns with NA.
}
\details{
BindFillRows
}
\examples{
df1 <- data.frame(a = seq_len(5), b = c(6:10))
df2 <- data.frame(a = c(11:15), b = c(16:20), c = LETTERS[seq_len(5)])
BindFillRows(df1, df2)
BindFillRows(list(df1, df2))

}
\keyword{internal}
