% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxBlur.R
\name{BoxBlur}
\alias{BoxBlur}
\title{Blur a matrix.}
\usage{
BoxBlur(mat.mtx, box.num = NULL, boxSize.num = NULL, sd.num = 1)
}
\arguments{
\item{mat.mtx}{<matrix>: Numerical matrix.}

\item{box.num}{<numeric>: The numerical vector for kernel. If NULL apply a GaussBox (see 'GaussBox' function) (Default NULL)}

\item{boxSize.num}{<numeric>: If box.num is NULL, size of kernel for 'GaussBox' function. (Default NULL)}

\item{sd.num}{<numeric>: If box.num is NULL, standard deviation parameter for 'GaussBox' function. (Default NULL)}
}
\value{
Blurred matrix.
}
\description{
Blur a matrix with a one dimensional kernel.
}
\details{
BoxBlur
}
\examples{
set.seed(981643)
mat.mtx <- rnorm(10000, 50, 10)**3 |> matrix(100, 100)
heatmap(mat.mtx, Rowv = NA, Colv = NA)
heatmap(BoxBlur(mat.mtx), Rowv = NA, Colv = NA)

}
\keyword{internal}
