% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakVector.R
\name{BreakVector}
\alias{BreakVector}
\title{Cut a vector.}
\usage{
BreakVector(
  x = NULL,
  x_min = NULL,
  center = NULL,
  x_max = NULL,
  n = 10,
  method = "linear"
)
}
\arguments{
\item{x}{<numeric>: Numerical vector.}

\item{x_min}{<numeric>: Minimal fixed value.}

\item{center}{<numeric>: Center fixed value.}

\item{x_max}{<numeric>: Maximal fixed value.}

\item{n}{<numeric>: Number of tile (return n+1 breaks).}

\item{method}{<character>: Kind of breaking. "linear" or "density". (Default "linear")}
}
\value{
Numerical vector of breaks.
}
\description{
Compute the n+1 breaks of a vector in a linear or density based way with the possibility to fix minimal, center and maximal values.
}
\details{
BreakVector
}
\examples{
set.seed(31415)
BreakVector(x = rnorm(100, 50, 200), n = 9)

}
\keyword{internal}
