% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreakVector.R
\name{BreakVector}
\alias{BreakVector}
\title{Cut a vector.}
\usage{
BreakVector(
  x.num = NULL,
  min.num = NULL,
  center.num = NULL,
  max.num = NULL,
  n.num = 10,
  method.chr = "linear"
)
}
\arguments{
\item{x.num}{<numeric>: Numerical vector.}

\item{min.num}{<numeric>: Minimal fixed value.}

\item{center.num}{<numeric>: Center fixed value.}

\item{max.num}{<numeric>: Maximal fixed value.}

\item{n.num}{<numeric>: Number of tile (return n.num+1 breaks).}

\item{method.chr}{<character>: Kind of breaking. "linear" or "density". (Default "linear")}
}
\value{
Numerical vector of breaks.
}
\description{
Compute the n+1 breaks of a vector in a linear or density based way with the possibility to fix minimal, center and maximal values.
}
\details{
BreakVector
}
\examples{
set.seed(31415)
BreakVector(x.num = rnorm(100, 50, 200), n.num = 9)

}
\keyword{internal}
