% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussBox.R
\name{GaussBox}
\alias{GaussBox}
\title{One dimension Gaussian kernel.}
\usage{
GaussBox(sd.num = 1, boxSize.num = NULL, scale.chr = "1")
}
\arguments{
\item{sd.num}{<numeric>: Standard deviation parameter of the gaussian. (Default 1)}

\item{boxSize.num}{<numeric>: Kernel size. If NULL size is 1+4*sd.num (Default NULL)}

\item{scale.chr}{<character>: Scaling kind of box. If "1" sum of kernel equal 1. If "int" Minimal value of kernel is 1 and all entry are integer. If "none", kernel is not scale. (Default "1")}
}
\value{
numerical vector.
}
\description{
One dimension Gaussian kernel.
}
\details{
GaussBox
}
\examples{
GaussBox(sd.num = 5, scale.chr = "none")
GaussBox(scale.chr = "1")
GaussBox(scale.chr = "int")

}
\keyword{internal}
