% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFileName.R
\name{GetFileName}
\alias{GetFileName}
\title{Get file name.}
\usage{
GetFileName(path.pth = NULL, ext.bln = FALSE)
}
\arguments{
\item{path.pth}{<character>: The path to the file.}

\item{ext.bln}{<logical>: Whether the file extension should be returned with the file name. (Default FALSE)}
}
\value{
A character string.
}
\description{
Function as `basename()` with the option to not return the file extension.
}
\details{
GetFileName
}
\examples{
filePath.pth ="my/path/to/my/file.txt"
GetFileName(path.pth=filePath.pth, ext.bln=FALSE)
GetFileName(path.pth=filePath.pth, ext.bln=TRUE)
}
\keyword{internal}
