% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndexFeatures.R
\name{IndexFeatures}
\alias{IndexFeatures}
\title{Indexes GRanges on genome.}
\usage{
IndexFeatures(
  gRangeList = NULL,
  genomicConstraint = NULL,
  chromSizes = NULL,
  binSize = NULL,
  method = "mean",
  metadataColName = NULL,
  cores = 1,
  verbose = FALSE
)
}
\arguments{
\item{gRangeList}{<GRanges or GRangesList or list[GRanges]>: GRanges object, list of GRanges or GRangesList containing coordinates to index.}

\item{genomicConstraint}{<GRanges>: GRanges object of constraint regions. Note that bins in the same constraint region only will be paired in HicAggR::SearchPairs(). If NULL chromosomes in chromSizes are used as constraints (Default NULL)}

\item{chromSizes}{<data.frame>: A data.frame containing chromosomes names and lengths in base pairs (see example).}

\item{binSize}{<integer>: Bin size in bp - corresponds to HiC matrix resolution.}

\item{method}{<character>: A string defining which summary method is used on metadata columns defined in metadataColName if multiple ranges are indexed in the same bin. Use 'mean', 'median', 'sum', 'max' or 'min'. (Default 'mean'')}

\item{metadataColName}{<character> : A character vector that specify the metadata columns of GRanges on which apply the summary method if multiple ranges are indexed in the same bin.}

\item{cores}{<integer> : Number of cores used. (Default 1)}

\item{verbose}{<logical>: If TRUE show the progression in console. (Default FALSE)}
}
\value{
A GRanges object.
}
\description{
Function that indexes a GRanges object on binned genome and constraints. Needed prior HicAggR::SearchPairs() function.
}
\details{
IndexFeatures
}
\examples{
data(Beaf32_Peaks.gnr)
Beaf32_Index.gnr <- IndexFeatures(
    gRangeList = list(Beaf = Beaf32_Peaks.gnr),
    chromSizes = data.frame(
        seqnames = c("2L", "2R"),
        seqlengths = c(23513712, 25286936)
    ),
    binSize = 100000
)

}
