% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IsRgb.R
\name{IsRgb}
\alias{IsRgb}
\title{Check RGB color format.}
\usage{
IsRgb(color.col = NULL)
}
\arguments{
\item{color.col}{<character or numeric>: A color.}
}
\value{
A logical.
}
\description{
Check if a color is in RGB color format.
}
\details{
IsRgb
}
\examples{
IsRgb("red")
IsRgb("#FFFFFF")
IsRgb(c(125, 125, 125))
IsRgb(c(43.8, 0.873, 0.492))

}
\keyword{internal}
