% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoinHiC.R
\name{JoinHiC}
\alias{JoinHiC}
\title{Merge HiC chunk.}
\usage{
JoinHiC(hic.cmx_lst)
}
\arguments{
\item{hic.cmx_lst}{<List[contactMatrix]>: The HiC maps list.}
}
\value{
A contactMatrix.
}
\description{
Create mega contactMatrix from a list of contactMatrix.
}
\details{
JoinHiC
}
\examples{
data(HiC_Ctrl.cmx_lst)
Mega_Ctrl.cmx <- JoinHiC(HiC_Ctrl.cmx_lst)

}
