% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{MinMaxScale}
\alias{MinMaxScale}
\title{Scales values on min-max range.}
\usage{
MinMaxScale(x, x_min = (0), x_max = 1)
}
\arguments{
\item{x}{\if{html}{\out{<numeric>}}: Numerical vector.}

\item{x_min}{\if{html}{\out{<numeric>}}: Minimal value after scaling.}

\item{x_max}{\if{html}{\out{<numeric>}}: Maximal value after scaling.}
}
\value{
Scaled numeric vector.
}
\description{
Scale values on min-max range.
}
\details{
MinMaxScale
}
\examples{
set.seed(655213)
x <- rnorm(500, 500)
set.seed(522613)
y.num <- rnorm(500, 100)
plot(density(x), col = "red", xlim = c(min(y.num), max(x)))
lines(density(y.num), col = "green")
plot(
    density(MinMaxScale(x)),
    col = "red",
    xlim = c(min(MinMaxScale(y.num)), max(MinMaxScale(x)))
)
lines(density(MinMaxScale(y.num)), col = "green")

}
\keyword{internal}
