% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OverExpectedHiC.R
\name{OverExpectedHiC}
\alias{OverExpectedHiC}
\title{Genomic distance bias correction.}
\usage{
OverExpectedHiC(hicLst, verbose = FALSE, cores = 1)
}
\arguments{
\item{hicLst}{<List[contactMatrix]>: The HiC maps list.}

\item{verbose}{<logical>: If TRUE show the progression in console. (Default FALSE)}

\item{cores}{<numerical> : Number of cores to be used. (Default 1)}
}
\value{
A matrices list.
}
\description{
Function that normalises HiC matrices by expected values computing by genomic distance.
}
\details{
OverExpectedHiC
}
\examples{
# Note: run HicAggR::BalanceHiC before OverExpectedHiC calculation.
data(HiC_Ctrl.cmx_lst)
OverExpectedHiC(HiC_Ctrl.cmx_lst)

}
