% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReduceRun.R
\name{ReduceRun}
\alias{ReduceRun}
\title{Apply a function over two RLE}
\usage{
ReduceRun(firstRle, secondRle, reduceMethod = "paste", ...)
}
\arguments{
\item{firstRle}{<rle or Rle>: First rle.}

\item{secondRle}{<rle or Rle>>: Second rle.}

\item{reduceMethod}{<character>: Name of a function to apply e.g paste, sum, mean.}

\item{...}{<...>: Other parameter for the reduce function.}
}
\value{
Reduced Rle
}
\description{
Apply a function on the values over two RLE and return one RLE.
}
\details{
ReduceRun
}
\examples{
firstRle <- rle(c("A", "A", "B"))
secondRle <- rle(c("A", "B", "B"))
ReduceRun(firstRle = firstRle, secondRle = secondRle, reduceMethod = "paste", sep = "_")
firstRle <- S4Vectors::Rle(c(1, 2, 3))
secondRle <- S4Vectors::Rle(c(5, 5, 5))
ReduceRun(firstRle = firstRle, secondRle = secondRle, reduceMethod = "sum")

}
\keyword{internal}
