% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReduceRun.R
\name{ReduceRun}
\alias{ReduceRun}
\title{Apply a function over two RLE}
\usage{
ReduceRun(first.rle, second.rle, reduceFun.chr = "paste", ...)
}
\arguments{
\item{first.rle}{<rle or Rle>: First rle.}

\item{second.rle}{<rle or Rle>>: Second rle.}

\item{reduceFun.chr}{<character>: Name of a function to apply e.g paste, sum, mean.}

\item{...}{<...>: Other parameter for the reduce function.}
}
\value{
Reduced Rle
}
\description{
Apply a function on the values over two RLE and return one RLE.
}
\details{
ReduceRun
}
\examples{
first.rle <- rle(c("A", "A", "B"))
second.rle <- rle(c("A", "B", "B"))
ReduceRun(first.rle = first.rle, second.rle = second.rle, reduceFun.chr = "paste", sep = "_")
first.rle <- S4Vectors::Rle(c(1, 2, 3))
second.rle <- S4Vectors::Rle(c(5, 5, 5))
ReduceRun(first.rle = first.rle, second.rle = second.rle, reduceFun.chr = "sum")

}
\keyword{internal}
