% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ResizeMatrix}
\alias{ResizeMatrix}
\title{Resize a matrix}
\usage{
ResizeMatrix(mtx, newDim = dim(mtx))
}
\arguments{
\item{mtx}{\if{html}{\out{<matrix>}}: A numerical matrix to resize.}

\item{newDim}{\if{html}{\out{<integer>}}: The number of rows and cols in
resized matrix.}
}
\value{
Resized matrix.
}
\description{
Resize a numericam matrix in new dimension.
}
\details{
ResizeMatrix
}
\keyword{internal}
