% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResizeMatrix.R
\name{ResizeMatrix}
\alias{ResizeMatrix}
\title{Resize a matrix}
\usage{
ResizeMatrix(matrice.mtx, newDim.num = dim(matrice.mtx))
}
\arguments{
\item{matrice.mtx}{<matrix>: A numerical matrix to resize.}

\item{newDim.num}{<integer>: The number of rows and cols in resized matrix.}
}
\value{
Resized matrix.
}
\description{
Resize a numericam matrix in new dimension.
}
\details{
ResizeMatrix
}
\examples{
matrice.mtx <- matrix(0, 11, 11)
matrice.mtx[which(as.logical(seq_len(11 * 11) \%\% 2))] <- seq_len(ceiling((11 * 11) / 2))
matrice.mtx[2, ] <- 100
matrice.mtx[, 7] <- 200
matrice.mtx
ResizeMatrix(matrice.mtx = matrice.mtx, newDim.num = c(7, 7))
ResizeMatrix(matrice.mtx = matrice.mtx, newDim.num = c(13, 13))
}
\keyword{internal}
