% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SdThreshold.R
\name{SdThreshold}
\alias{SdThreshold}
\title{Find threshold for outliers based on sd.}
\usage{
SdThreshold(x.num = NULL, sdThreshold.num = 3, bounds.chr = "both")
}
\arguments{
\item{x.num}{<numeric>: numeric vector.}

\item{sdThreshold.num}{<numeric>: number of standard deviation. (Default 3)}

\item{bounds.chr}{<character>: bounds to return, "lower", "upper" or "both". (Default "both")}
}
\value{
numerical vector of thresholds values for outliers triming
}
\description{
Find threshold for outliers triming based on standard deviation.
}
\details{
SdThreshold
}
\examples{
set.seed(1111)
x.num <- rnorm(1000)
x.num <- sort(x.num)
x.num
SdThreshold(x.num, sdThreshold.num = 2, bounds.chr = "lower")
SdThreshold(x.num, sdThreshold.num = 2, bounds.chr = "both")
SdThreshold(x.num, sdThreshold.num = 2, bounds.chr = "upper")

}
\keyword{internal}
