% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SearchPairs.R
\name{SearchPairs}
\alias{SearchPairs}
\title{Creates pairs from genomic index.}
\usage{
SearchPairs(
  indexAnchor.gnr = NULL,
  indexBait.gnr = NULL,
  minDist.num = NULL,
  maxDist.num = NULL,
  verbose.bln = FALSE,
  cores.num = 1
)
}
\arguments{
\item{indexAnchor.gnr}{<GRanges>: A first indexed GRanges object used as pairs anchor (must be indexed using IndexFeatures()).}

\item{indexBait.gnr}{<GRanges>: A second indexed GRanges object used as pairs bait (must be indexed using IndexFeatures()). If NULL, indexAnchor.gnr is used instead (Default NULL)}

\item{minDist.num}{<numeric>: Minimal distance between anchors and baits. (Default NULL)}

\item{maxDist.num}{<numeric>: Maximal distance between anchors and baits. (Default NULL)}

\item{verbose.bln}{<logical>: If TRUE show the progression in console. (Default FALSE)}

\item{cores.num}{<integer> : Number of cores to use. (Default 1)}
}
\value{
A GInteractions object.
}
\description{
Creates pairs of coordinates from indexed anchor and bait genomic coordinates according to distance constraints.
}
\details{
SearchPairs
}
\examples{
# Data
data(Beaf32_Peaks.gnr)

# Index Beaf32
Beaf32_Index.gnr <- IndexFeatures(
    gRange.gnr_lst = list(Beaf = Beaf32_Peaks.gnr),
    chromSize.dtf = data.frame(seqnames = c("2L", "2R"), seqlengths = c(23513712, 25286936)),
    binSize.num = 100000
)

# Beaf32 <-> Beaf32 Pairing
Beaf_Beaf.gni <- SearchPairs(indexAnchor.gnr = Beaf32_Index.gnr)

}
