% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WrapFunction.R
\name{WrapFunction}
\alias{WrapFunction}
\title{Convert string to function.}
\usage{
WrapFunction(...)
}
\arguments{
\item{...}{<character>: A string that could be parse and eval as a function.}
}
\value{
The result of the function or a function.
}
\description{
Wrap a string into a function.
}
\details{
WrapFunction
}
\examples{
X <- c(2, 4, NA)
myString <- "mean(X, na.rm=TRUE)"
WrapFunction(myString)
myResult <- WrapFunction(myString)
myString_2 <- "function(X){mean(X, na.rm=TRUE)}"
WrapFunction(myString_2)
myFunction <- WrapFunction(myString_2)
myFunction(c(1, 2, 3))

}
\keyword{internal}
